# docs/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Kumbukum API",
    "version": "1.0.0",
    "description": "Everything your AI needs to remember — Notes, Memory, URLs, AI Chat"
  },
  "servers": [
    {
      "url": "https://your-instance.example.com/api/v1",
      "description": "API v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token from login"
      },
      "AccessToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Personal access token. Value: `Token <your-token>`"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
    },
...
  },
  "security": [
    {
      "BearerAuth": []
    },
    {
      "AccessToken": []
    }
  ],
  "paths": {
...
    "/screenshots/{filename}": {
      "get": {
        "tags": [
          "Screenshots"
        ],
        "summary": "Get a screenshot image (signed URL)",
        "description": "Serves a screenshot image. Requires valid HMAC signature and unexpired timestamp. No authentication needed.",
        "security": [],
        "parameters": [
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Screenshot filename (SHA-256 hash + .png)"
          },
          {
            "name": "expires",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Expiration timestamp (Unix epoch seconds)"
          },
          {
            "name": "sig",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "HMAC-SHA256 signature"
          }
        ],
        "responses": {
          "200": {
            "description": "PNG image",
            "content": {
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "403": {
            "description": "Invalid or expired signature",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Screenshot not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}